p/test-case/requirements.txt
google-generativeai>=0.1.0
Pillow>=9.0.0
python-dotenv>=0.19.0